requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==2.4.1
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.2.0
blessings==1.7
blinker==1.8.2
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
idna==3.6
import-ipynb==0.1.4
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
packaging==23.2
pafy==0.5.5
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plumbum==1.8.3
pooch==1.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-mon==2.0.5
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==73.0.1
six==1.16.0
soundfile==0.12.1
soxr==0.4.0
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.4.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
youtube-dl==2021.12.17
